+++ pkg/frontend/querymiddleware/split_and_cache.go
@@ -625,7 +625,7 @@ func splitQueryByInterval(r Request, interval time.Duration) ([]Request, error) 
 func evaluateAtModifierFunction(query string, start, end int64) (string, error) {
 	expr, err := parser.ParseExpr(query)
 	if err != nil {
-		return "", apierror.New(apierror.TypeBadData, err.Error())
+		return "", apierror.New(apierror.TypeBadData, decorateWithParamName(err, "query").Error())
 	}
 	parser.Inspect(expr, func(n parser.Node, _ []parser.Node) error {
 		if selector, ok := n.(*parser.VectorSelector); ok {