+++ pkg/querier/timeseries_series_set_test.go
@@ -8,6 +8,7 @@ package querier
 import (
 	"testing"
 
+	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/tsdb/chunkenc"
 	"github.com/stretchr/testify/require"
 
@@ -44,8 +45,7 @@ func TestTimeSeriesSeriesSet(t *testing.T) {
 	require.True(t, ss.Next())
 	series := ss.At()
 
-	require.Equal(t, ss.ts[0].Labels[0].Name, series.Labels()[0].Name)
-	require.Equal(t, ss.ts[0].Labels[0].Value, series.Labels()[0].Value)
+	require.Equal(t, labels.FromStrings("label1", "value1"), series.Labels())
 
 	it := series.Iterator(nil)
 	require.Equal(t, chunkenc.ValFloat, it.Next())